+++ frontend/src/pages/Committees.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 
 import Card from '../components/Card'
-import LoadingSpinner from '../components/Loader'
+import LoadingSpinner from '../components/LoadingSpinner'
 import Pagination from '../components/Pagination'
 import SearchBar from '../components/Search'
 import { loadData } from '../lib/api'
